# docs/swagger.yaml
basePath: /api
definitions:
...
  entity.Task:
    properties:
      description:
        type: string
      id:
        type: integer
      is_done:
        type: boolean
      name:
        type: string
      user_id:
        type: integer
    type: object
  entity.TaskCreate:
    properties:
      description:
        type: string
      is_done:
        type: boolean
      name:
        type: string
    required:
    - description
    - name
    type: object
  entity.TaskUpdate:
    properties:
      description:
        type: string
      is_done:
        type: boolean
      name:
        type: string
    type: object
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: API Server for TodoApp
  title: ToDoListApi
  version: "1.0"
paths:
...
  /api/tasks:
    post:
      consumes:
      - application/json
      description: Create task
      operationId: create-task
      parameters:
      - description: task body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.TaskCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Task
      tags:
      - tasks
  /api/tasks/:id:
    delete:
      consumes:
      - application/json
      description: Delete Task
      operationId: delete-task
      parameters:
      - description: task id
        in: path
        name: input
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Task
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Get Task Info
      operationId: get-task-info
      parameters:
      - description: task id
        in: path
        name: input
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Task Info
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Update Task Info
      operationId: update-task-info
      parameters:
      - description: task id
        in: path
        name: input
        required: true
        type: integer
      - description: updated task info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.TaskUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Task Info
      tags:
      - tasks
  /api/tasks/export:
    get:
      description: Export Task
      operationId: Export-task
      produces:
      - text/csv
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Export Task
      tags:
      - tasks
  /api/tasks/import:
    post:
      consumes:
      - multipart/form-data
      description: Import Task
      operationId: import-task
      parameters:
      - description: CSV file with tasks
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Import Task
      tags:
      - tasks
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"